+++ src/app.module.ts
@@ -8,7 +8,6 @@ import { KakaoLoginModule } from './kakao/login/kakao-login.module';
 import { OpenAiModule } from './open-ai/open-ai.module';
 import { PrismaModule } from './prisma/prisma.module';
 import { ProductModule } from './product/product.module';
-import { ProductReviewModule } from './product/review/product-review.module';
 import { ReviewModule } from './review/review.module';
 import { S3Module } from './s3/s3.module';
 import { SearchModule } from './search/search.module';
@@ -27,7 +26,6 @@ import { WishlistProductModule } from './wishlist/product/wishlist-product.modul
     OpenAiModule,
     CrawlerModule,
     TasksModule,
-    ProductReviewModule,
     ProductModule,
     SearchModule,
     WishlistProductModule,
