+++ src/review/review.controller.spec.ts
@@ -6,10 +6,12 @@ import { DeepMockProxy, mockDeep } from 'jest-mock-extended';
 
 import { REVIEWS_PAGE_SIZE } from '@/constants/review';
 import { PrismaService } from '@/prisma/prisma.service';
-import { FindReviewsDto } from '@/product/review/dto/find-reviews.dto';
-import { ReviewCountDto } from '@/review/dto/review-count.dto';
-import { fileMock, reviewMock, reviewMockWithImgs } from '@/utils/unit-test';
+import { ReviewsWithReviwerDto } from '@/review/dto/reviews-with-reviewer.dto';
+import { fileMock, mockUser, reviewMock, reviewMockWithImgs } from '@/utils/unit-test';
 
+import { CreateReviewDto } from './dto/create-review.dto';
+import { FindReviewsDto } from './dto/find-reviews.dto';
+import { ReviewCountDto } from './dto/review-count.dto';
 import { ReviewDto } from './dto/review.dto';
 import { ReviewsDto } from './dto/reviews.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
@@ -17,7 +19,7 @@ import { ReviewController } from './review.controller';
 import { ReviewService } from './review.service';
 
 import { TokenGuardReq } from '@/types/refreshTokenGuard.type';
-import { OrderBy, ReviewsData, SortBy } from '@/types/review.type';
+import { OrderBy, ReviewsData, ReviewsWithReviewerData, SortBy } from '@/types/review.type';
 
 describe('ReviewController', () => {
   let controller: ReviewController;
@@ -43,6 +45,18 @@ describe('ReviewController', () => {
     expect(controller).toBeDefined();
   });
 
+  describe('create', () => {
+    it('should return an instance of ReviewDto', async () => {
+      const req = {
+        id: reviewMock.userId
+      } as TokenGuardReq;
+      const body: CreateReviewDto = { text: reviewMock.text, rating: reviewMock.rating };
+      service.create.mockResolvedValue(reviewMockWithImgs);
+      const result = await controller.create(req, body, reviewMock.productId, []);
+      expect(result).toBeInstanceOf(ReviewDto);
+    });
+  });
+
   describe('findOne', () => {
     it('should return an instance of ReviewDto', async () => {
       service.findOne.mockResolvedValue(reviewMockWithImgs);
@@ -51,7 +65,30 @@ describe('ReviewController', () => {
     });
   });
 
-  describe('findMany', () => {
+  describe('findProductReviews', () => {
+    it('should return an instance of ReviewsDto', async () => {
+      const sortBy: SortBy = 'createdAt';
+      const orderBy: OrderBy = 'desc';
+      const page = 1;
+      const reviewMocksWithReviewerData: ReviewsWithReviewerData = {
+        reviews: [
+          {
+            ...reviewMockWithImgs,
+            reviewer: mockUser
+          }
+        ],
+        pageSize: REVIEWS_PAGE_SIZE,
+        total: 1,
+        totalPages: 1
+      };
+      service.findProductReviews.mockResolvedValue(reviewMocksWithReviewerData);
+      const getReviewsDto: FindReviewsDto = { sortBy, orderBy, page };
+      const result = await controller.findProductReviews(reviewMock.productId, getReviewsDto);
+      expect(result).toBeInstanceOf(ReviewsWithReviwerDto);
+    });
+  });
+
+  describe('findUserReviews', () => {
     it('should return an instance of ReviewsDto', async () => {
       const req = {
         id: reviewMock.userId
@@ -65,9 +102,9 @@ describe('ReviewController', () => {
         total: 1,
         totalPages: 1
       };
-      service.findMany.mockResolvedValue(reviewsData);
+      service.findUserReviews.mockResolvedValue(reviewsData);
       const getReviewsDto: FindReviewsDto = { sortBy, orderBy, page };
-      const result = await controller.findMany(req, getReviewsDto);
+      const result = await controller.findUserReviews(req, getReviewsDto);
       expect(result).toBeInstanceOf(ReviewsDto);
     });
   });
