+++ src/review/review.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   ParseIntPipe,
   Patch,
+  Post,
   Query,
   Req,
   UnsupportedMediaTypeException,
@@ -19,10 +20,12 @@ import { plainToInstance } from 'class-transformer';
 
 import { AccessTokenGuard } from '@/auth/guard/access-token.guard';
 import { reviewImgInterceptor } from '@/interceptor/review.interceptor';
-import { FindReviewsDto } from '@/product/review/dto/find-reviews.dto';
-import { ReviewCountDto } from '@/review/dto/review-count.dto';
+import { ReviewsWithReviwerDto } from '@/review/dto/reviews-with-reviewer.dto';
 import { isValidFileContent } from '@/utils/file';
 
+import { CreateReviewDto } from './dto/create-review.dto';
+import { FindReviewsDto } from './dto/find-reviews.dto';
+import { ReviewCountDto } from './dto/review-count.dto';
 import { ReviewDto } from './dto/review.dto';
 import { ReviewsDto } from './dto/reviews.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
@@ -30,10 +33,44 @@ import { ReviewService } from './review.service';
 
 import { TokenGuardReq } from '@/types/refreshTokenGuard.type';
 
-@Controller('reviews')
+@Controller('')
 @ApiTags('review')
 export class ReviewController {
   constructor(private readonly service: ReviewService) {}
+
+  @ApiOperation({
+    summary: 'Create a product review'
+  })
+  @ApiResponse({
+    status: 201,
+    description: 'Created',
+    type: ReviewDto
+  })
+  @ApiResponse({
+    status: 400,
+    description: 'Bad Request - Invalid or missing fields in the request body'
+  })
+  @ApiResponse({
+    status: 401,
+    description: 'Unauthorized - Invalid or expired access token, or access token is missing'
+  })
+  @ApiResponse({
+    status: 409,
+    description: 'Conflict - User has already submitted a review for this product'
+  })
+  @ApiConsumes('multipart/form-data')
+  @UseGuards(AccessTokenGuard)
+  @UseInterceptors(reviewImgInterceptor)
+  @Post('products/:id/reviews')
+  async create(
+    @Req() { id }: TokenGuardReq,
+    @Body() createReviewDto: CreateReviewDto,
+    @Param('id', ParseIntPipe) productId: number,
+    @UploadedFiles() imgs?: Express.Multer.File[]
+  ) {
+    const review = await this.service.create({ createReviewDto, userId: id, productId, imgs });
+    return plainToInstance(ReviewDto, review);
+  }
 
   @ApiOperation({
     summary: 'Get product review count'
@@ -48,14 +85,14 @@ export class ReviewController {
     description: 'Unauthorized - Invalid or expired access token, or access token is missing'
   })
   @UseGuards(AccessTokenGuard)
-  @Get('count')
+  @Get('reviews/count')
   async count(@Req() { id }: TokenGuardReq) {
     const count = await this.service.count(id);
     return plainToInstance(ReviewCountDto, count);
   }
 
   @ApiOperation({
-    summary: 'Get product reviews'
+    summary: 'Get user reviews'
   })
   @ApiResponse({
     status: 200,
@@ -71,10 +108,35 @@ export class ReviewController {
     description: 'Unauthorized - Invalid or expired access token, or access token is missing'
   })
   @UseGuards(AccessTokenGuard)
-  @Get('')
-  async findMany(@Req() { id }: TokenGuardReq, @Query() findReviewsDto: FindReviewsDto) {
-    const reviews = await this.service.findMany(findReviewsDto, id);
+  @Get('reviews')
+  async findUserReviews(@Req() { id }: TokenGuardReq, @Query() findReviewsDto: FindReviewsDto) {
+    const reviews = await this.service.findUserReviews(findReviewsDto, id);
     return plainToInstance(ReviewsDto, reviews);
+  }
+
+  @ApiOperation({
+    summary: 'Get product reviews'
+  })
+  @ApiResponse({
+    status: 200,
+    description: 'Success',
+    type: ReviewsWithReviwerDto
+  })
+  @ApiResponse({
+    status: 400,
+    description: 'Bad Request - Invalid or missing query fields'
+  })
+  @ApiResponse({
+    status: 404,
+    description: 'Not Found - Product not found'
+  })
+  @Get('products/:id/reviews')
+  async findProductReviews(
+    @Param('id', ParseIntPipe) productId: number,
+    @Query() findReviewsDto: FindReviewsDto
+  ) {
+    const reviews = await this.service.findProductReviews(findReviewsDto, productId);
+    return plainToInstance(ReviewsWithReviwerDto, reviews);
   }
 
   @ApiOperation({
@@ -89,7 +151,7 @@ export class ReviewController {
     status: 404,
     description: 'Not Found - Review not found'
   })
-  @Get(':id')
+  @Get('reviews/:id')
   async findOne(@Param('id', ParseIntPipe) reviewId: number) {
     const review = await this.service.findOne(reviewId);
     return plainToInstance(ReviewDto, review);
@@ -130,7 +192,7 @@ export class ReviewController {
   @ApiConsumes('multipart/form-data')
   @UseGuards(AccessTokenGuard)
   @UseInterceptors(reviewImgInterceptor)
-  @Patch(':id')
+  @Patch('reviews/:id')
   async update(
     @Req() req: TokenGuardReq,
     @Body() updateReviewDto: UpdateReviewDto,
@@ -175,7 +237,7 @@ export class ReviewController {
   })
   @UseGuards(AccessTokenGuard)
   @HttpCode(204)
-  @Delete(':id')
+  @Delete('reviews/:id')
   async remove(@Req() { id }: TokenGuardReq, @Param('id', ParseIntPipe) reviewId: number) {
     await this.service.remove(id, reviewId);
   }
