+++ test/controller/review.integration-spec.ts
@@ -3,12 +3,17 @@ import { Product } from '@prisma/client';
 import * as request from 'supertest';
 
 import { AuthModule } from '@/auth/auth.module';
+import { KaKaoUserInfoDto } from '@/kakao/login/dto/kakao-user-info.dto';
+import { KakaoLoginService } from '@/kakao/login/kakao-login.service';
 import { PrismaService } from '@/prisma/prisma.service';
+import { CreateReviewDto } from '@/review/dto/create-review.dto';
 import { ReviewCountDto } from '@/review/dto/review-count.dto';
 import { ReviewDto } from '@/review/dto/review.dto';
+import { ReviewsWithReviwerDto } from '@/review/dto/reviews-with-reviewer.dto';
 import { ReviewsDto } from '@/review/dto/reviews.dto';
 import { UpdateReviewDto } from '@/review/dto/update-review.dto';
 import { ReviewModule } from '@/review/review.module';
+import { ReviewService } from '@/review/review.service';
 import {
   createProduct,
   createReview,
@@ -19,9 +24,11 @@ import {
   validateDto
 } from '@/utils/integration-test';
 
-describe('ProductReviewController (integration)', () => {
+describe('ReviewController (integration)', () => {
   let app: INestApplication;
   let prismaService: PrismaService;
+  let kakaoLoginService: KakaoLoginService;
+  let reviewService: ReviewService;
 
   const userId = '2';
   const anotherUserId = '3';
@@ -32,6 +39,8 @@ describe('ProductReviewController (integration)', () => {
   beforeAll(async () => {
     app = await createTestingApp([ReviewModule, AuthModule]);
     prismaService = app.get(PrismaService);
+    kakaoLoginService = app.get(KakaoLoginService);
+    reviewService = app.get(ReviewService);
 
     const user = await createUser(app, userId);
     const anotherUser = await createUser(app, anotherUserId);
@@ -44,6 +53,64 @@ describe('ProductReviewController (integration)', () => {
     await deleteUser(prismaService, userId);
     await deleteUser(prismaService, anotherUserId);
     await deleteProduct(prismaService, product.id);
+  });
+
+  describe('/api/products/:id/reviews (POST)', () => {
+    it('201', async () => {
+      const req: CreateReviewDto = {
+        text: 'Great product!',
+        rating: 5
+      };
+      const { body } = await request(app.getHttpServer())
+        .post(`/api/products/${product.id}/reviews`)
+        .set('Cookie', [accessTokenCookie])
+        .field('text', req.text)
+        .field('rating', req.rating)
+        .attach('imgs', 'test/test-img.png')
+        .expect(201);
+      validateDto(ReviewDto, body);
+
+      const { id } = await prismaService.review.findUnique({
+        where: { productId_userId: { productId: product.id, userId } },
+        select: { id: true }
+      });
+
+      await reviewService.remove(userId, id);
+    });
+
+    it('400', async () => {
+      await request(app.getHttpServer())
+        .post('/api/products/1/reviews')
+        .set('Cookie', [accessTokenCookie])
+        .send({ text: 'Great product!', rating: 0 })
+        .expect(400);
+    });
+
+    it('401', async () => {
+      return request(app.getHttpServer()).post('/api/products/1/reviews').expect(401);
+    });
+
+    it('404', async () => {
+      await request(app.getHttpServer())
+        .post('/api/products/0/reviews')
+        .set('Cookie', [accessTokenCookie])
+        .send({ text: 'Great product!', rating: 5 })
+        .expect(404);
+    });
+
+    it('409', async () => {
+      await createReview(prismaService, { userId, productId: product.id });
+
+      await request(app.getHttpServer())
+        .post(`/api/products/${product.id}/reviews`)
+        .set('Cookie', [accessTokenCookie])
+        .send({ text: 'Great product!', rating: 5 })
+        .expect(409);
+
+      await prismaService.review.delete({
+        where: { productId_userId: { productId: product.id, userId } }
+      });
+    });
   });
 
   describe('/api/reviews/:id (GET)', () => {
@@ -84,6 +151,33 @@ describe('ProductReviewController (integration)', () => {
 
     it('401', async () => {
       await request(app.getHttpServer()).get('/api/reviews?page=1').expect(401);
+    });
+  });
+
+  describe('/api/products/:id/reviews (GET)', () => {
+    it('200', async () => {
+      await createReview(prismaService, { userId, productId: product.id });
+
+      const kakaoUsers: KaKaoUserInfoDto[] = [
+        { id: userId, nickname: 'nickname', profileImg: 'url' }
+      ];
+      kakaoLoginService.findUsers = jest.fn().mockResolvedValue(kakaoUsers);
+      const { body } = await request(app.getHttpServer())
+        .get(`/api/products/${product.id}/reviews?page=1`)
+        .expect(200);
+      validateDto(ReviewsWithReviwerDto, body);
+
+      await prismaService.review.delete({
+        where: { productId_userId: { productId: product.id, userId } }
+      });
+    });
+
+    it('400', async () => {
+      await request(app.getHttpServer()).get('/api/products/0/reviews?page=a').expect(400);
+    });
+
+    it('404', async () => {
+      await request(app.getHttpServer()).get('/api/products/0/reviews?page=1').expect(404);
     });
   });
 
@@ -179,31 +273,17 @@ describe('ProductReviewController (integration)', () => {
     });
 
     it('415 - fake PNG file', async () => {
-      const patchReviewDto: UpdateReviewDto = {
-        text: 'new-text',
-        rating: 5
-      };
-
       await request(app.getHttpServer())
         .patch('/api/reviews/0')
-        .field('text', patchReviewDto.text)
-        .field('rating', patchReviewDto.rating)
         .attach('imgs', 'test/fake-png.png')
         .set('Cookie', [accessTokenCookie])
         .expect(415);
     });
 
     it('415', async () => {
-      const patchReviewDto: UpdateReviewDto = {
-        text: 'new-text',
-        rating: 5
-      };
-
       await request(app.getHttpServer())
         .patch('/api/reviews/0')
-        .field('text', patchReviewDto.text)
-        .field('rating', patchReviewDto.rating)
-        .attach('imgs', 'test/unsupported.pdf')
+        .attach('imgs', 'test/unsupported.md')
         .set('Cookie', [accessTokenCookie])
         .expect(415);
     });
